// apps/messenger/src/models/profile.model.ts
export interface status {
  away_message?: string;
  away_message_mode: string;
  is_signed_out: boolean;
  r?: [];
  r_type?: string[];
  s?: string;
  c?: {
    ref: string;
    type: string;
  };
  role_notify?:ROLE_NOTIFY[];
}
// ...
export interface ProfileStatus {
  firstAndLastName: string;
  status?: string;
}